requirements.txt
blessings==1.6
bpython==0.14.1
curtsies==0.1.19
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.1
django-allauth==0.20.0
django-braces==1.8.1
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
greenlet==0.4.7
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6.1
Pygments==2.0.2
PyMySQL==0.6.6
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
